+++ pkgs/crane-utils/default.nix
@@ -5,8 +5,8 @@
 rustPlatform.buildRustPackage {
   pname = "crane-utils";
   version = "0.0.1";
+  useFetchCargoVendor = true;
 
   src = lib.sourceFilesBySuffices ./. [ ".rs" ".toml" ".lock" ];
-
-  cargoHash = "sha256-yVe6BIUuZygz44GiPRWAjMCU9IbglyrS2RAuLcP+3Ls=";
+  cargoLock = { lockFile = ./Cargo.lock; };
 }